Update tests
Add tests for PyQt6 and PySide6.
Fix PySide2 test failing due to `pytest-qt` dropped its support in 4.5.0.
For Qt5 bindings, Python 3.7 (EOL @ 2023-06-27) remains in the tests, for it mostly works still. I would bump the Python version up as soon as a failure occurs, without wasting time on a fix.